feat: add party tile widget
